Allow loading voxel-engine, as a synthetic "master plugin"
The game instance passed to voxel-plugins' constructor can now be null,
and voxel-engine loaded as if it were a "plugin" instead. A special plugin,
however. All plugins automatically depend on it, its instance is set to
this.game for passing to all other plugins, and when instantiated,
only the options are passed to its constructor. Mainly useful for allowing
voxel-engine's options to be configured along with other plugins.